在庫管理～～～～～～～～ンゴ～～～～～

+++ src/main/java/jp/ken/jdbc/domain/mapper/GoodsRowMapper.java
@@ -16,10 +16,10 @@ public class GoodsRowMapper implements RowMapper<GoodsEntity> {
 
         GoodsEntity g = new GoodsEntity();
 
-        g.setGoodsId(rs.getLong("goods_id"));
+        g.setGoodsId(rs.getInt("goods_id"));
         g.setGoodsName(rs.getString("goods_name"));
-        g.setCategoryId(rs.getLong("category_id"));
-        g.setGenreId(rs.getLong("genre_id"));
+        g.setCategoryId(rs.getInt("category_id"));
+        g.setGenreId(rs.getInt("genre_id"));
         g.setQuantity(rs.getInt("quantity"));
         g.setJanCode(rs.getString("jan_code"));
         g.setImageUrl(rs.getString("image_url"));
